Перенес MainActivity во фрагмент
