+++ app/src/main/java/com/example/note_taker_project/UI/NoteListFragment.java
@@ -1,0 +1,70 @@
+package com.example.note_taker_project.UI;
+
+import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.Button;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+import androidx.recyclerview.widget.ItemTouchHelper;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
+
+import com.example.note_taker_project.App;
+import com.example.note_taker_project.Domain.Note;
+import com.example.note_taker_project.Domain.NoteRepository;
+import com.example.note_taker_project.R;
+
+public class NoteListFragment extends Fragment {
+    private RecyclerView recyclerView;
+    private NoteAdapter adapter;
+    private NoteRepository noteRepository;
+    private NoteListener noteListener;
+
+    private Button addButton;
+
+    @Nullable
+    @Override
+    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
+        return inflater.inflate(R.layout.fragment_notes_list, container, false);
+    }
+
+    @Override
+    public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
+        noteListener = new NoteListener(getContext());
+        initRecycler(view);
+        addButton = view.findViewById(R.id.main_activity_add_button);
+        addButton.setOnClickListener(v -> {
+            noteListener.onAddNote();
+        });
+    }
+
+    private void initRecycler(View view) {
+        recyclerView = view.findViewById(R.id.recycler_view);
+        recyclerView.setLayoutManager(new LinearLayoutManager(getContext()));
+        adapter = App.get().adapter;
+        noteRepository = App.get().noteRepository;
+        adapter.setData(noteRepository.getNotes());
+        adapter.setOnClickListener(noteListener);
+        new ItemTouchHelper(itemTouchHelperCallback).attachToRecyclerView(recyclerView);
+        recyclerView.setAdapter(adapter);
+    }
+
+    private final ItemTouchHelper.SimpleCallback itemTouchHelperCallback = new ItemTouchHelper.SimpleCallback(0, ItemTouchHelper.RIGHT) {
+        @Override
+        public boolean onMove(@NonNull RecyclerView recyclerView, @NonNull RecyclerView.ViewHolder viewHolder, @NonNull RecyclerView.ViewHolder target) {
+            return false;
+        }
+
+        @Override
+        public void onSwiped(@NonNull RecyclerView.ViewHolder viewHolder, int i) {
+            int index = viewHolder.getAdapterPosition();
+            Note note = noteRepository.getNotes().get(index);
+            noteRepository.deleteNote(note);
+            adapter.deleteItem(note);
+        }
+    };
+}
