+++ app/src/main/java/com/example/note_taker_project/UI/NoteListener.java
@@ -1,5 +1,6 @@
 package com.example.note_taker_project.UI;
 
+import android.content.Context;
 import android.content.Intent;
 
 import com.example.note_taker_project.App;
@@ -7,10 +8,10 @@ import com.example.note_taker_project.Domain.Note;
 
 class NoteListener implements OnNoteListener {
 
-    private final MainActivity mainActivity;
+    private final Context context;
 
-    public NoteListener(MainActivity mainActivity) {
-        this.mainActivity = mainActivity;
+    public NoteListener(Context context) {
+        this.context = context;
     }
 
     @Override
@@ -21,15 +22,15 @@ class NoteListener implements OnNoteListener {
 
     @Override
     public void onClickNote(Note note) {
-        Intent intent = new Intent(mainActivity, InfoItemNoteActivity.class);
+        Intent intent = new Intent(context, InfoItemNoteActivity.class);
         intent.putExtra(InfoItemNoteActivity.NOTE_EXTRA_KEY, note);
-        mainActivity.startActivity(intent);
+        context.startActivity(intent);
     }
 
     @Override
     public void onAddNote() {
-        Intent intent = new Intent(mainActivity, AddItemNoteActivity.class);
-        mainActivity.startActivity(intent);
+        Intent intent = new Intent(context, AddItemNoteActivity.class);
+        context.startActivity(intent);
     }
 
 }
